+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 - [set reporter.enabled default to true](https://github.com/Tencent/spring-cloud-tencent/pull/786)
 - [fix:fix nacos enabled condition bugs.](https://github.com/Tencent/spring-cloud-tencent/pull/794)
 - [fix:fix instance's weight hard code to 100 bug.](https://github.com/Tencent/spring-cloud-tencent/pull/801)
+- [feat:update getPort method of PolarisRegistration.](https://github.com/Tencent/spring-cloud-tencent/pull/804)